+++ server/middleware/removeDemoUserChanges.js
@@ -1,7 +1,4 @@
-const List = require('../models/list.model');
-const Cohort = require('../models/cohort.model');
-const User = require('../models/user.model');
-const Comment = require('../models/comment.model');
+const { removeDemoUserData } = require('../common/utils/userUtils');
 
 const removeDemoUserChanges = async (req, res, next) => {
   if (!req.user) {
@@ -12,36 +9,10 @@ const removeDemoUserChanges = async (req, res, next) => {
 
   if (idFromProvider === process.env.DEMO_USER_ID_FROM_PROVIDER) {
     try {
-      const lists = await List.find(
-        {
-          $or: [
-            { ownerIds: currentUserId },
-            { memberIds: currentUserId },
-            { viewersIds: currentUserId }
-          ]
-        },
-        '_id'
-      )
-        .lean()
-        .exec();
+      await removeDemoUserData(currentUserId);
 
-      if (lists) {
-        const listsIds = lists.map(lists => lists._id);
-        await Comment.deleteMany({ listId: { $in: listsIds } });
-      }
-      await List.deleteMany({
-        $or: [
-          { ownerIds: currentUserId },
-          { memberIds: currentUserId },
-          { viewersIds: currentUserId }
-        ]
-      }).exec();
-      await Cohort.deleteMany({
-        $or: [{ ownerIds: currentUserId }, { memberIds: currentUserId }]
-      }).exec();
-      await User.deleteOne({ _id: currentUserId });
-      await User.deleteMany({ provider: `demo-${currentUserId}` });
-    } catch {
+      return next();
+    } catch (err) {
       return res
         .status(400)
         .send({ message: 'Logout failed. Please try again.' });
